fix: add filter isFile() for 04-copy-directory

+++ 04-copy-directory/index.js
@@ -12,14 +12,24 @@ const copyDir = () => {
 			})
 		}
 		fs.mkdir(pathCopy, () => {
-			fs.readdir(join(__dirname, 'files'), (err, files) => {
-				files.forEach((file) => {
-					fsPromises.copyFile(
-						join(__dirname, 'files', file),
-						join(__dirname, 'files-copy', file)
-					)
-				})
-			})
+			fs.readdir(
+				join(__dirname, 'files'),
+				{ withFileTypes: true },
+				(err, files) => {
+					files.forEach((file) => {
+						if (file.isFile()) {
+							fsPromises.copyFile(
+								join(__dirname, 'files', file.name),
+								join(__dirname, 'files-copy', file.name)
+							)
+						} else {
+							console.log(
+								'В папку files добавлена папка. В задании нет требования копировать папки и рекурсивно копировать их вложения'
+							)
+						}
+					})
+				}
+			)
 		})
 	})
 }
